refactor: move auth custom classes to auth namespace
