+++ app/Http/Controllers/AuthController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Components\AuthHandler;
-use App\Components\ForgotPasswordHandler;
+use App\Components\Auth\AuthHandler;
+use App\Components\Auth\ForgotPasswordHandler;
 use App\Http\Requests\Login;
 use App\Http\Requests\ResetPassword;
 use App\Http\Requests\ResetRequest;
